requirements.txt
pandas
scikit-learn
lightgbm
streamlit
google-generativeai
jupyterlab
tensorflow
sentence-transformers
numpy
notebook
tf-keras
